norminette parsing and work on expansion
