+++ src/ast/tokenization/helper_f.c
@@ -1,0 +1,34 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   helper_f.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/10/17 19:38:06 by jhor              #+#    #+#             */
+/*   Updated: 2025/10/17 19:38:35 by jhor             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../../minishell.h"
+
+void	free_tokens(t_token *tokens)
+{
+	t_token	*temp;
+
+	while (tokens != NULL)
+	{
+		temp = tokens;
+		tokens = tokens->next;
+		free(temp->lexeme);
+		free(temp);
+	}
+	free(tokens);
+}
+
+char	*trim_prompt(char *trim)
+{
+	while (*trim == ' ' || *trim == '\t')
+		trim++;
+	return (trim);
+}
