+++ src/create_node.c
@@ -6,13 +6,13 @@
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/14 17:50:35 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/15 15:02:34 by jhor             ###   ########.fr       */
+/*   Updated: 2025/10/17 16:48:12 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-t_token	*create_node(t_token *head, char *start, size_t end) //!fix the space problem in double quotes
+t_token	*create_node(t_token *head, char *start, size_t end)
 {
 	t_token	*temp;
 	size_t		i;
@@ -34,19 +34,16 @@ t_token	*create_node(t_token *head, char *start, size_t end) //!fix the space pr
 	temp->lexeme[i] = '\0';
 	temp->next = NULL;
 	head = temp;
-	// printf("create node is: %s\n", head->lexeme);
-	// printf("0000000000000000000000000000000000000\n");
 	return (head);
 }
 
 t_token	*append_node(t_token *head, char *start, size_t end)
 {
 	t_token	*temp;
-	size_t		i;
+	size_t	i;
+	t_token	*temp2;
+	t_token	*trave;
 	
-	// printf("iterator in append: %zu\n", end);
-	// for(size_t j = 0; j < end; j++)
-	// 	printf("%c\n", start[j]);
 	i = 0;
 	temp = malloc(sizeof(t_token));
 	if (!temp)
@@ -56,12 +53,9 @@ t_token	*append_node(t_token *head, char *start, size_t end)
 	temp->lexeme = malloc(sizeof(char) * (end + 1));
 	if (!temp->lexeme)
 		return (NULL);
-	t_token *temp2 = NULL;
+	temp2 = NULL;
 	if (head->next != NULL)
 		temp2 = head->next;
-	printf("head node is null or not: %s\n", head->lexeme);
-	if (temp2 != NULL)
-		printf("second node is null or not: %s\n", temp2->lexeme);
 	while (i < end && start[i])
 	{
 		temp->lexeme[i] = start[i];
@@ -69,19 +63,19 @@ t_token	*append_node(t_token *head, char *start, size_t end)
 	}
 	temp->lexeme[i] = '\0';
 	temp->next = NULL;
-	t_token *trave = NULL;
+	trave = NULL;
 	trave = head;
 	while (trave->next != NULL)
 		trave = trave->next;
 	trave->next = temp;
-	t_token *temp3 = NULL;
-	temp3 = trave->next;
-	while (temp3 != NULL)
-	{
-		printf("new node: %s\n", temp3->lexeme);
-		temp3 = temp3->next;
-	}
-	printf("----------------------------------------------\n");
+	// t_token *temp3 = NULL;
+	// temp3 = trave->next;
+	// while (temp3 != NULL)
+	// {
+	// 	printf("new node: %s\n", temp3->lexeme);
+	// 	temp3 = temp3->next;
+	// }
+	// printf("----------------------------------------------\n");
 	return (head);
 }
 
@@ -89,27 +83,16 @@ t_token	*append_word_node(t_token *head, char *start, size_t n_start, size_t end
 {
 	t_token	*temp;
 	size_t	i;
+	t_token	*trave;
 	
 	i = 0;
-	// printf("start iterator in append: %zu\n", n_start);
-	// printf("end iterator in append: %zu\n", end);
-	// for(size_t j = 0; j < end; j++)
-	// 	printf("%c\n", start[j]);
 	temp = NULL;
 	temp = malloc(sizeof(t_token));
-	if (!temp)
-		return (NULL);
 	temp->token = 0;
 	temp->lexeme = NULL;
 	temp->lexeme = malloc(sizeof(char) * (end + 1));
 	if (!temp->lexeme)
 		return (NULL);
-	// t_token *temp2 = NULL;
-	// if (head->next != NULL)
-	// 	temp2 = head->next;
-	// printf("head node is null or not: %s\n", head->lexeme);
-	// if (temp2 != NULL)
-		// printf("second node is null or not: %s\n", temp2->lexeme);
 	while (n_start < end && start[n_start])
 	{
 		temp->lexeme[i++] = start[n_start];
@@ -117,20 +100,10 @@ t_token	*append_word_node(t_token *head, char *start, size_t n_start, size_t end
 	}
 	temp->lexeme[i] = '\0';
 	temp->next = NULL;
-	
-	t_token *trave = NULL;
+	trave = NULL;
 	trave = head;
 	while (trave->next != NULL)
 		trave = trave->next;
 	trave->next = temp;
-	
-	// t_token *temp3 = NULL;
-	// temp3 = temp;
-	// while (temp3 != NULL)
-	// {
-	// 	printf("new node: %s\n", temp3->lexeme);
-	// 		temp3 = temp3->next;
-	// }
-	// printf("----------------------------------------------\n");
 	return (head);
 }
